requirements.txt
asn1crypto==1.5.1
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
DateTime==5.1
docutils==0.16
feedparser==6.0.10
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Session==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
numpy==1.24.3
oscrypto==1.3.0
packaging==23.1
protobuf==3.20.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pycryptodomex==3.18.0
PyJWT==2.7.0
pyodbc==4.0.39
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
requests==2.31.0
requests-file==1.5.1
rsa==4.7.2
sgmllib3k==1.0.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tldextract==3.4.4
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.4
zope.interface==6.0
